Clearly separate infrastructure code (implementation detal) from domain logic
