+++ src/main/scala/io/github/pauljamescleary/petstore/domain/model/Order.scala
@@ -1,0 +1,11 @@
+package io.github.pauljamescleary.petstore.domain.model
+
+import java.time.Instant
+
+case class Order(
+    petId: Long,
+    shipDate: Option[Instant] = None,
+    status: OrderStatus = Placed,
+    complete: Boolean = false,
+    id: Option[Long] = None
+)
