--- src/main/scala/io/github/pauljamescleary/petstore/model/User.scala
@@ -1,11 +1,0 @@
-package io.github.pauljamescleary.petstore.model
-
-case class User(
-    userName: String,
-    firstName: String,
-    lastName: String,
-    email: String,
-    password: String,
-    phone: String,
-    id: Option[Long] = None
-)
